# Remove assembly scanning from DynamoDBEntryConversion