## Description
Replace assembly scanning in DynamoDBEntryConversion which finds converter types with explicit types created.

## Motivation and Context
Assembly scanning has a cold start performance impact, which is especially important for serverless applications. In this case, it scans own assembly for types, which doesn't have much sense, as the results of that scan is static, it will never change. Therefore, it doesn't make sense to find those types in runtime.

Besides, it is more AOT-friendly, and it probably allows to remove most of RequiresUnreferencedCode attributes (not sure).

## Testing

I have created a small benchmark to test cold start performance.

<details>
  <summary>Benchmark</summary>
  
  ```csharp
    [SimpleJob(RunStrategy.ColdStart, launchCount: 20, iterationCount: 1, invocationCount: 1)]
    [MinColumn, MaxColumn, MeanColumn, MedianColumn]
    public class ConversionBenchmark
    {
        [Benchmark]
        public void CreateConversions()
        {
            _ = DynamoDBEntryConversion.V1;
            _ = DynamoDBEntryConversion.V2;
        }
    }
  ```
  
</details>


Before:
| Method            | Mean     | Error    | StdDev   | Min      | Max      | Median   |
|------------------ |---------:|---------:|---------:|---------:|---------:|---------:|
| CreateConversions | 44.38 ms | 4.562 ms | 5.253 ms | 33.70 ms | 53.47 ms | 44.96 ms |7 ms |

After:
| Method            | Mean     | Error    | StdDev   | Min      | Max      | Median   |
|------------------ |---------:|---------:|---------:|---------:|---------:|---------:|
| CreateConversions | 14.87 ms | 2.421 ms | 2.788 ms | 10.34 ms | 22.29 ms | 14.33 ms |


## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist
<!--- Go over all the following points, and put an `x` in all the boxes that apply -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project
- [ ] My change requires a change to the documentation
- [ ] I have updated the documentation accordingly
- [x] I have read the **README** document
- [ ] I have added tests to cover my changes
- [x] All new and existing tests passed

## License
<!--- The SDK is released under the [Apache 2.0 license][license], so any code you submit will be released under that license -->
<!--- For substantial contributions, we may ask you to sign a [Contributor License Agreement (CLA)][cla] -->
<!--- Put an `x` in the below box if you confirm that this request can be released under the Apache 2 license -->
- [x] I confirm that this pull request can be released under the Apache 2 license

[issues]: https://github.com/aws/aws-sdk-net/issues
[license]: http://aws.amazon.com/apache2.0/
[cla]: http://en.wikipedia.org/wiki/Contributor_License_Agreement