@normj 2024-04-04
I like this idea of getting rid of the Assembly scan. This was written at a time predating Lambda when cold starts weren't so critical. We need a way to make sure if somebody adds a new converter it doesn't get forgotten to be added here. I'm thinking a simple approach is add a `ConverterCount` property to `DynamoDBEntryConversion`. Then you could add a unit test to checks the known number and then run the assembly scan in their now to compute the actual number for the unit test.