@Dreamescaper 2024-04-05
@normj 
I have added a unit test verifying that all converters are registered (had to add InternalsVisibleTo attribute).

(Not sure I like it though, focusing on the implementation details, but probably better than nothing).