@normj 2024-04-24
The PR went out today as part of version `3.7.302.22` of AWSSDK.DynamoDBv2. Thanks for the PR and helping improve the SDK's performance.